Change repository id type to Integer

+++ internship-warehouse/src/main/java/com/twoday/internshipwarehouse/services/ProductService.java
@@ -29,7 +29,7 @@ public class ProductService {
             throw new InvalidValueException(Constants.QUANTITY);
         }
 
-        Product product = productRepository.findById((long) productSellRequest.id()).orElse(null);
+        Product product = productRepository.findById(productSellRequest.id()).orElse(null);
 
         if (product == null) {
             throw new ProductNotFoundByIdException(productSellRequest.id());
